[Refactor] UIFont, UIColor extension 추가
