+++ Kindy/Kindy/Assets.xcassets/Colors/AccentColor.colorset/Contents.json
