+++ Kindy/Kindy/Assets.xcassets/Colors/kindyLightGray.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xE4",
-          "green" : "0xE4",
-          "red" : "0xE4"
+          "blue" : "0.894",
+          "green" : "0.894",
+          "red" : "0.894"
         }
       },
       "idiom" : "universal"
