+++ Kindy/Kindy/Assets.xcassets/Colors/kindySecondaryLightGreen.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0x57",
-          "green" : "0x74",
-          "red" : "0x25"
+          "blue" : "205",
+          "green" : "223",
+          "red" : "209"
         }
       },
       "idiom" : "universal"
