+++ Kindy/Kindy/Extensions/UIColor+.swift
@@ -8,5 +8,9 @@
 import UIKit
 
 extension UIColor {
-    
+    static let kindyGray = UIColor(named: "kindyGray")
+    static let kindyLightGray = UIColor(named: "kindyLightGray")
+    static let kindyPrimaryGreen = UIColor(named: "kindyPrimaryGreen")
+    static let kindySecondaryGreen = UIColor(named: "kindySecondaryGreen")
+    static let kindySecondaryLightGreen = UIColor(named: "kindySecondaryLightGreen ")
 }
