+++ Kindy/Kindy/Extensions/UIFont+.swift
@@ -1,0 +1,19 @@
+//
+//  UIFont+.swift
+//  Kindy
+//
+//  Created by 정호윤 on 2022/10/30.
+//
+
+import UIKit
+
+extension UIFont {
+    static let title1 = UIFont(name: "AppleSDGothicNeo-SemiBold", size: 28)
+    static let title2 = UIFont(name: "AppleSDGothicNeo-Bold", size: 24)
+    static let title3 = UIFont(name: "AppleSDGothicNeo-Regular", size: 20)
+    static let headline = UIFont(name: "AppleSDGothicNeo-Bold", size: 17)
+    static let body1 = UIFont(name: "AppleSDGothicNeo-Regular", size: 17)
+    static let body2 = UIFont(name: "AppleSDGothicNeo-Regular", size: 15)
+    static let subhead = UIFont(name: "AppleSDGothicNeo-Bold", size: 15)
+    static let footnote = UIFont(name: "AppleSDGothicNeo-Regular", size: 13)
+}
