+++ Kindy/Kindy/Extensions/UIImage+.swift
@@ -12,10 +12,9 @@ extension UIImage {
     func resizeImage(size: CGSize) -> UIImage {
         let originalSize = self.size
         
-        let ratio: CGFloat = {
-            return originalSize.width > originalSize.height ? 1 / (size.width / originalSize.width) :
-            1 / (size.height / originalSize.height)
-        }()
+        var ratio: CGFloat {
+            return originalSize.width > originalSize.height ? 1 / (size.width / originalSize.width) : 1 / (size.height / originalSize.height)
+        }
         
         return UIImage(cgImage: self.cgImage!, scale: self.scale * ratio, orientation: self.imageOrientation)
     }
