+++ Kindy/Kindy/Models/Bookstore/Bookstore.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 // 서점 구조체(임시)
-struct Bookstore: Hashable {
+struct Bookstore {
     
     let id = UUID()
     
@@ -38,14 +38,6 @@ struct Bookstore: Hashable {
         return 1
     }
     
-    static func == (lhs: Bookstore, rhs: Bookstore) -> Bool {
-        return lhs.id == rhs.id
-    }
-    
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-    
     static let dummyData: [Bookstore] = [
         Bookstore(images: nil, name: "달팽이 책방1", address: "포항시 남구", telNumber: "020202020", emailAddress: "teamsandalsofficial@gmail.com", instagramURL: nil, businessHour: BusinessHour(), description: "내 손 안의 독립서점, 킨디", location: Location(latitude: 10, longitude: 10)),
         Bookstore(images: nil, name: "달팽이 책방2", address: "포항시 남구", telNumber: "020202020", emailAddress: "teamsandalsofficial@gmail.com", instagramURL: nil, businessHour: BusinessHour(), description: "내 손 안의 독립서점, 킨디", location: Location(latitude: 10, longitude: 10)),
@@ -53,23 +45,15 @@ struct Bookstore: Hashable {
         Bookstore(images: nil, name: "달팽이 책방4", address: "포항시 남구", telNumber: "020202020", emailAddress: "teamsandalsofficial@gmail.com", instagramURL: nil, businessHour: BusinessHour(), description: "내 손 안의 독립서점, 킨디", location: Location(latitude: 10, longitude: 10)),
         Bookstore(images: nil, name: "달팽이 책방5", address: "포항시 남구", telNumber: "020202020", emailAddress: "teamsandalsofficial@gmail.com", instagramURL: nil, businessHour: BusinessHour(), description: "내 손 안의 독립서점, 킨디", location: Location(latitude: 10, longitude: 10))
     ]
+}
+
+extension Bookstore: Hashable {
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(id)
+    }
     
-}
-
-struct BusinessHour {
-    var monday: String = ""
-    var tuesday: String = ""
-    var wednesday: String = ""
-    var thursday: String = ""
-    var friday: String = ""
-    var saturday: String = ""
-    var sunday: String = ""
-    var etc: String? = ""
-}
-
-// 서점의 위도와 경도를 표현하는 구조체
-struct Location {
-    let latitude: Double
-    let longitude: Double
+    static func == (lhs: Bookstore, rhs: Bookstore) -> Bool {
+        return lhs.id == rhs.id
+    }
 }
