+++ Kindy/Kindy/Models/Bookstore/BusinessHour.swift
@@ -1,0 +1,19 @@
+//
+//  BusinessHour.swift
+//  Kindy
+//
+//  Created by 정호윤 on 2022/10/30.
+//
+
+import Foundation
+
+struct BusinessHour {
+    var monday: String = ""
+    var tuesday: String = ""
+    var wednesday: String = ""
+    var thursday: String = ""
+    var friday: String = ""
+    var saturday: String = ""
+    var sunday: String = ""
+    var etc: String? = ""
+}
