+++ Kindy/Kindy/View Controllers/Curation/PagingCurationViewController.swift
