+++ Kindy/Kindy/View Controllers/My Page/DetailMyPageViewController.swift
